\name{LLCT-package}
\alias{LLCT-package}
\docType{package}
\title{
\packageTitle{LLCT}
}
\description{
\packageDescription{LLCT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LLCT}
\packageIndices{LLCT}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{LLCT}

Maintainer: \packageMaintainer{LLCT}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
